--- logs/benchmark.txt
@@ -1,96 +1,0 @@
-Single Pendulum:
-
-Algorithm 1 (Explicit)
-
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.009131847909000498
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.007913920890001465
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.007325533929000812
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.007358514274998015
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.008577905581001686
-
-=> 0.008061544516800495
-
-Algorithm 0 (Implicit)
-
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.016354156733999845
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.015790636683999763
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.010209442785999272
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.011113003326998296
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.00960726074899867
-
-=> 0.00960726074899867
-
-Algorithm Exp (Hybrid)
-
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.009085318935998657
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.009965997180002887
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.008378236066000681
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.00784329078600058
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.007801849158000551
-
-=> 0.00861493842520067
-
-=========================================
-
-Double Pendulum:
-
-Algorithm 1 (Explicit)
-
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.49753417840998737
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.44546206916998926
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.40896136226001545
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.44720536144002837
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.4071949527099787
-
-=> 0.4412715847979999
-
-Algorithm 0 (Implicit)
-
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py 
-0.44198834549999444
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.4142982527899949
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.36426657502001036
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.3691325602300276
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % /Users/linussommer/miniforge3/envs/sage/bin/python /Users/linussommer/Documents/GitHub/pendulum/core/main.py
-0.38353316323999026
-
-=> 0.39367058231400004
-
-Algorithm Exp (Hybrid)
-
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % python core/main.py 
-0.36112922674998116
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % python core/main.py
-0.3648366582900053
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % python core/main.py
-0.3666759101599746
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % python core/main.py
-0.35722548612997346
-(sage) (base) linussommer@x86_64-apple-darwin13 pendulum % python core/main.py
-0.38829160485998726
-
-=> 0.36763177723798435
-
