+++ systems/hard/kepler_half.py
